+++ src/models/error.model.ts
@@ -1,0 +1,5 @@
+export default interface IError {
+    title: string
+    status: number
+    message: string
+}