+++ src/tests/App.test.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { render, screen } from '@testing-library/react';
+import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { store } from '../app/store';
 import App from '../App';
 
-test('Renders ', () => {
+test('Renders the App with name', () => {
   const { getByText } = render(
     <Provider store={store}>
       <App />