+++ src/tests/fetch.test.ts
@@ -1,16 +1,16 @@
-import { fetchHandler } from '../app/fetch'
+import fetchHandler from '../app/fetch'
 
 test('gets mojombo from list of users', async () => {
-  const fetchedData = await fetchHandler('https://api.github.com/users')
+  const fetchedData = await fetchHandler('users')
   expect(fetchedData[0].login).toBe('mojombo')
 });
 
 test('gets mojombo as single user', async () => {
-  const fetchedData = await fetchHandler('https://api.github.com/users/mojombo')
+  const fetchedData = await fetchHandler('users/mojombo')
   expect(fetchedData.login).toBe('mojombo')
 });
 
 test('returns error for wrong URL', async () => {
-  const fetchedData = await fetchHandler('https://api.github.com/userz')
+  const fetchedData = await fetchHandler('userz')
   expect(fetchedData.error.status).toBe(404)
 });